:construction: Alterar o login no google drive, agora é na tela de login. :rocket:  Deletar funcionário da sessão. :rocket:  Criar pastas de imagens  ao configurar dispositivo.

+++ app/src/main/java/com/br/minasfrango/data/dao/ConfiguracaoGoogleDriveDAO.java
@@ -13,7 +13,7 @@ public class ConfiguracaoGoogleDriveDAO extends GenericsDAO<ConfiguracaoGoogleDr
         return new ConfiguracaoGoogleDriveDAO(type);
     }
 
-    public ConfiguracaoGoogleDrive pesquisarPorIdDoFuncionario(final int idUsuario) {
+    public ConfiguracaoGoogleDrive pesquisarPorIdDoFuncionario(final long idUsuario) {
 
         ConfiguracaoGoogleDriveORM configuracaoGoogleDriveORM = where().equalTo("idFuncionario",idUsuario).findFirst();
         return new ConfiguracaoGoogleDrive(configuracaoGoogleDriveORM);
